When double click and call new nodes, should not draw new nods but need to add new nodes.
