--- src/components/PresentationComponents/NetworkGraph/NetworkDiagramSVGUpdateNewNodes.tsx
@@ -1,136 +1,0 @@
-import React, { useEffect, useRef } from 'react';
-import * as d3 from 'd3';
-
-const App: React.FC = () => {
-    const svgRef = useRef<SVGSVGElement | null>(null);
-    const width = 500;
-    const height = 300;
-
-    let graph: { nodes: any[]; links: any[] } = { nodes: [], links: [] };
-
-    const randomizeData = () => {
-        let n = Math.floor(Math.random() * 10) + 6;
-
-        let newNodes = [];
-        for (let i = 0; i < n; i++) {
-            if (graph.nodes[i]) newNodes.push(graph.nodes[i]);
-            else
-                newNodes.push({
-                    id: i,
-                    color: Math.floor(Math.random() * 10),
-                    size: Math.floor(Math.random() * 10 + 2),
-                    x: Math.random() * width,
-                    y: Math.random() * height,
-                });
-        }
-
-        let newLinks = [];
-        let m = Math.floor(Math.random() * n) + 1;
-        for (let i = 0; i < m; i++) {
-            let a = 0;
-            let b = 0;
-            while (a === b) {
-                a = Math.floor(Math.random() * n);
-                b = Math.floor(Math.random() * n);
-            }
-            newLinks.push({ source: a, target: b });
-            if (i < newNodes.length - 2) newLinks.push({ source: i, target: i + 1 });
-        }
-
-        graph = { nodes: newNodes, links: newLinks };
-        update();
-    };
-
-    useEffect(() => {
-        randomizeData();
-    }, []);
-
-    let links: d3.Selection<SVGLineElement, any, SVGGElement, unknown>;
-    let nodes: d3.Selection<SVGGElement, any, SVGGElement, unknown>;
-
-    const update = () => {
-        const svg = d3.select<SVGSVGElement, unknown>(svgRef.current as any);
-
-        nodes = svg.selectAll<SVGGElement, unknown>('g').data(graph.nodes as any[]);
-
-        nodes.exit().remove();
-
-        const newNodes = nodes.enter().append('g').attr('opacity', 0);
-
-        newNodes
-            .transition()
-            .attr('opacity', 1)
-            .attr('class', 'nodes')
-            .attr('r', (d: any) => (d.size * 2) + 1)
-            .attr('fill', (d: any) => d3.schemeCategory10[d.color]);
-
-        newNodes
-            .append('text')
-            .text((d: any) => d.id)
-            .attr('x', 6)
-            .attr('y', 3)
-            .style('font-size', '20px');
-
-        newNodes.append('title').text((d: any) => d.id);
-
-        nodes = newNodes.merge(nodes);
-        nodes.append('circle')
-            .attr('r', (d: any) => (d.size * 2) + 1)
-            .attr('fill', (d: any) => d3.schemeCategory10[d.color]);
-
-        links = svg.selectAll<SVGLineElement, unknown>('line').data(graph.links as any[]);
-
-        links.exit().remove();
-
-        const newLinks = links.enter().append('line').attr('stroke-width', (d: any) => Math.sqrt(d.value));
-
-        newLinks.attr('opacity', 0).transition().attr('opacity', 1);
-
-        links = newLinks.merge(links);
-
-        const simulation = d3.forceSimulation(graph.nodes);
-
-        simulation.force(
-            'link',
-            d3.forceLink(graph.links).id((d: any) => d.id).strength(0.004)
-        );
-
-        simulation.force('charge', d3.forceManyBody());
-        simulation.force('x', d3.forceX(width / 2).strength(0.01));
-        simulation.force('y', d3.forceY(height / 2).strength(0.01));
-
-        simulation.on('tick', ticked);
-        simulation.alpha(1).restart();
-    };
-
-    const ticked = () => {
-        links
-            .attr('x1', (d: any) => d.source.x)
-            .attr('y1', (d: any) => d.source.y)
-            .attr('x2', (d: any) => d.target.x)
-            .attr('y2', (d: any) => d.target.y);
-
-        nodes.attr('transform', (d: any) => `translate(${d.x},${d.y})`);
-    };
-
-    return (
-        <div>
-            <button onClick={randomizeData}>Update</button>
-            <svg ref={svgRef} width={width} height={height}></svg>
-            <style>
-                {`
-        .links line {
-          stroke: #999;
-          stroke-opacity: 0.6;
-        }
-        .nodes circle {
-          stroke: #fff;
-          stroke-width: 1.5px;
-        }
-      `}
-            </style>
-        </div>
-    );
-};
-
-export default App;
